+++ pyccel/codegen/pipeline.py
@@ -20,6 +20,7 @@ from pyccel.parser.parser          import Parser
 from pyccel.codegen.codegen        import Codegen
 from pyccel.codegen.utilities      import recompile_object
 from pyccel.codegen.utilities      import copy_internal_library
+from pyccel.codegen.utilities      import generate_extension_modules
 from pyccel.codegen.utilities      import internal_libs
 from pyccel.codegen.utilities      import external_libs
 from pyccel.codegen.python_wrapper import create_shared_library
@@ -338,10 +339,23 @@ def execute_pyccel(fname, *,
 
     # Iterate over the external_libs list and determine if the printer
     # requires an external lib to be included.
-    for key in codegen.get_printer_imports():
-        lib_name = key.split("/", 1)[0]
-        if lib_name in external_libs:
-            lib_dest_path = copy_internal_library(lib_name, pyccel_dirpath)
+    for key, import_node in codegen.get_printer_imports().items():
+        try:
+            deps = generate_extension_modules(key, import_node, pyccel_dirpath, language)
+        except NotImplementedError as error:
+            errors.report(f'{error}\n'+PYCCEL_RESTRICTION_TODO,
+                severity='error',
+                traceback=error.__traceback__)
+            handle_error('code generation (wrapping)')
+            raise PyccelCodegenError(msg) from None
+        except PyccelError:
+            handle_error('code generation (wrapping)')
+            raise
+        for d in deps:
+            recompile_object(d,
+                              compiler = src_compiler,
+                              verbose  = verbose)
+            mod_obj.add_dependencies(d)
 
     if convert_only:
         # Change working directory back to starting point