+++ pyccel/codegen/utilities.py
@@ -14,6 +14,7 @@ from filelock import FileLock
 import pyccel.stdlib as stdlib_folder
 import pyccel.extensions as ext_folder
 
+from .codegen              import printer_registry
 from .compiling.basic      import CompileObj
 from .compiling.file_locks import FileLockSet
 
@@ -30,7 +31,7 @@ language_extension = {'fortran':'f90', 'c':'c', 'python':'py'}
 
 #==============================================================================
 # map external libraries inside pyccel/extensions with their path
-external_libs = {"stc"  : "STC/include",}
+external_libs = {"stc" : "STC/include/stc", "gFTL" : "gFTL/install/GFTL-1.13/include/v2"}
 
 #==============================================================================
 # map internal libraries to their folders inside pyccel/stdlib and their compile objects
@@ -114,7 +115,7 @@ def copy_internal_library(lib_folder, pyccel_dirpath, extra_files = None):
     """
     # get lib path (stdlib_path/lib_name or ext_path/lib_name)
     if lib_folder in external_libs:
-        lib_path = os.path.join(ext_path, external_libs[lib_folder], lib_folder)
+        lib_path = os.path.join(ext_path, external_libs[lib_folder])
     else:
         lib_path = os.path.join(stdlib_path, lib_folder)
 
@@ -177,6 +178,55 @@ def copy_internal_library(lib_folder, pyccel_dirpath, extra_files = None):
                         with open(extra_file, 'w', encoding="utf-8") as f:
                             f.writelines(contents)
     return lib_dest_path
+
+#==============================================================================
+def generate_extension_modules(import_key, import_node, pyccel_dirpath, language):
+    """
+    Generate any new modules that describe extensions.
+
+    Generate any new modules that describe extensions. This is the case for lists/
+    sets/dicts/etc handled by gFTL.
+
+    Parameters
+    ----------
+    import_key : str
+        The name by which the extension is identified in the import.
+    import_node : Import
+        The import used in the code generator (this object contains the module to
+        be printed).
+    pyccel_dirpath : str
+        The folder where files are being saved.
+    language : str
+        The language in which code is being printed.
+
+    Returns
+    -------
+    list[CompileObj]
+        A list of any new compilation dependencies which are required to compile
+        the translated file.
+    """
+    new_dependencies = []
+    lib_name = import_key.split('/', 1)[0]
+    if lib_name == 'gFTL_extensions':
+        lib_name = 'gFTL'
+        mod = import_node.source_module
+        printer = printer_registry[language]
+        filename = os.path.join(pyccel_dirpath, import_key)+'.F90'
+        folder = os.path.dirname(filename)
+        code = printer(filename).doprint(mod)
+        if not os.path.exists(folder):
+            os.mkdir(folder)
+        with FileLock(f'{folder}.lock'):
+            with open(filename, 'w', encoding="utf-8") as f:
+                f.write(code)
+
+        new_dependencies.append(CompileObj(os.path.basename(filename), folder=folder,
+                            includes=(os.path.join(pyccel_dirpath, 'gFTL'),)))
+
+    if lib_name in external_libs:
+        copy_internal_library(lib_name, pyccel_dirpath)
+
+    return new_dependencies
 
 #==============================================================================
 def recompile_object(compile_obj,