+++ tests/epyccel/test_epyccel_variable_annotations.py
@@ -217,55 +217,66 @@ def test_homogeneous_set_annotation_complex(stc_language):
     assert epyc_homogeneous_set_annotation() == homogeneous_set_annotation()
     assert isinstance(epyc_homogeneous_set_annotation(), type(homogeneous_set_annotation()))
 
-def test_homogeneous_empty_list_annotation_int(stc_language):
+def test_homogeneous_empty_list_annotation_int(language):
     def homogeneous_list_annotation():
         a: 'list[int]'  # pylint: disable=unused-variable
         a = []
-    epyc_homogeneous_list_annotation = epyccel(homogeneous_list_annotation, language=stc_language)
+    epyc_homogeneous_list_annotation = epyccel(homogeneous_list_annotation, language=language)
     assert epyc_homogeneous_list_annotation() == homogeneous_list_annotation()
     assert isinstance(epyc_homogeneous_list_annotation(), type(homogeneous_list_annotation()))
 
-def test_homogeneous_list_annotation_int(stc_language):
+def test_homogeneous_list_annotation_int(language):
     def homogeneous_list_annotation():
         a: 'list[int]'  # pylint: disable=unused-variable
         a = [1, 2, 3, 4]
         return a[0], a[1], a[2], a[3]
-    epyc_homogeneous_list_annotation = epyccel(homogeneous_list_annotation, language=stc_language)
+    epyc_homogeneous_list_annotation = epyccel(homogeneous_list_annotation, language=language)
     assert epyc_homogeneous_list_annotation() == homogeneous_list_annotation()
     assert isinstance(epyc_homogeneous_list_annotation(), type(homogeneous_list_annotation()))
 
-def test_homogeneous_list_without_annotation(stc_language):
+def test_homogeneous_list_without_annotation(language):
     def homogeneous_list():
         a = [1, 2, 3, 4] # pylint: disable=unused-variable
         return a[0], a[1], a[2], a[3]
-    epyc_homogeneous_list = epyccel(homogeneous_list, language=stc_language)
+    epyc_homogeneous_list = epyccel(homogeneous_list, language=language)
     assert epyc_homogeneous_list() == homogeneous_list()
     assert isinstance(epyc_homogeneous_list(), type(homogeneous_list()))
 
-def test_homogeneous_list_annotation_float(stc_language):
+def test_homogeneous_list_annotation_float(language):
     def homogeneous_list_annotation():
         a: 'list[float]'  # pylint: disable=unused-variable
         a = [1.1, 2.2, 3.3, 4.4]
         return a[0], a[1], a[2], a[3]
-    epyc_homogeneous_list_annotation = epyccel(homogeneous_list_annotation, language=stc_language)
+    epyc_homogeneous_list_annotation = epyccel(homogeneous_list_annotation, language=language)
     assert epyc_homogeneous_list_annotation() == homogeneous_list_annotation()
     assert isinstance(epyc_homogeneous_list_annotation(), type(homogeneous_list_annotation()))
 
-def test_homogeneous_list_annotation_bool(stc_language):
+def test_homogeneous_list_annotation_float64(language):
+    def homogeneous_list_annotation():
+        from numpy import float64
+        a: 'list[float64]'  # pylint: disable=unused-variable
+        a = [1.1, 2.2, 3.3, 4.4]
+        return a[0], a[1], a[2], a[3]
+    epyc_homogeneous_list_annotation = epyccel(homogeneous_list_annotation, language=language)
+    assert epyc_homogeneous_list_annotation() == homogeneous_list_annotation()
+    assert isinstance(epyc_homogeneous_list_annotation(), type(homogeneous_list_annotation()))
+    assert isinstance(epyc_homogeneous_list_annotation()[0], type(homogeneous_list_annotation()[0]))
+
+def test_homogeneous_list_annotation_bool(language):
     def homogeneous_list_annotation():
         a: 'list[bool]'  # pylint: disable=unused-variable
         a = [False, True, True, False]
         return a[0], a[1], a[2], a[3]
-    epyc_homogeneous_list_annotation = epyccel(homogeneous_list_annotation, language=stc_language)
+    epyc_homogeneous_list_annotation = epyccel(homogeneous_list_annotation, language=language)
     assert epyc_homogeneous_list_annotation() == homogeneous_list_annotation()
     assert isinstance(epyc_homogeneous_list_annotation(), type(homogeneous_list_annotation()))
 
-def test_homogeneous_list_annotation_complex(stc_language):
+def test_homogeneous_list_annotation_complex(language):
     def homogeneous_list_annotation():
         a: 'list[complex]'  # pylint: disable=unused-variable
         a = [1+1j, 2+2j, 3+3j, 4+4j]
         return a[0], a[1], a[2], a[3]
-    epyc_homogeneous_list_annotation = epyccel(homogeneous_list_annotation, language=stc_language)
+    epyc_homogeneous_list_annotation = epyccel(homogeneous_list_annotation, language=language)
     assert epyc_homogeneous_list_annotation() == homogeneous_list_annotation()
     assert isinstance(epyc_homogeneous_list_annotation(), type(homogeneous_list_annotation()))